backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
